refine(resource_table.utils.PrepareResourceHelper): improve bundled resources processing efficiency and robustness; truly enable OTA download resume  (#16)
This PR introduces refinements to the resource_table.utils.PrepareResourceHelper as follow:
1. bundled resources processing: improve the robustness and efficiency on bundle resource broken, now bundled resources processing can recover itself properly from resource broken.
2. bundled resources processing: improve bundle downloading logic, now waiting thread workers can take over the downloading soon after the previous downloading thread failed, instead of just waiting.
3. bundled resources processing: extracted resources will also be hashed to confirm the integrity.
4. OTA download resume and re-use previous OTA download intermediate files are implemented in PrepareResourceHelper now.
5. ResumeOTADownloadHelper is moved to ota-image-libs due to it depends on ota-image-libs.utils.PrepareResourceHelper's implementation details.

This PR will bump ota-image-libs to v0.2.4.

+++ src/ota_image_libs/v1/resource_table/utils.py
@@ -15,32 +15,67 @@
 
 from __future__ import annotations
 
-import logging
+import itertools
 import os
+import shutil
 import threading
+import time
+from concurrent.futures import ThreadPoolExecutor
 from dataclasses import dataclass
+from hashlib import sha256
 from pathlib import Path
 from typing import Generator, Optional
 
 import zstandard
+from typing_extensions import TypeAlias
 
 from ota_image_libs._resource_filter import BundleFilter, CompressFilter, SliceFilter
 from ota_image_libs.common import tmp_fname
+from ota_image_libs.common.io import file_sha256, remove_file
 
-from .db import ResourceTableORMPool
+from .db import ResourceTableDBHelper, ResourceTableORMPool
 from .schema import ResourceTableManifest
 
-logger = logging.getLogger(__name__)
+# totally a worker thread will wait for 18 seconds
+#   for the bundle to be prepared.
+MAX_ITERS_FOR_WAITING_BUNDLE = 6  # seconds
+WAITING_BUNDLE_INTERVAL = 3
 
 
-def recreate_bundled_resource(
-    entry: ResourceTableManifest, bundle_fpath: Path, save_dst: Path
-) -> None:
-    filter_cfg = entry.filter_applied
-    assert isinstance(filter_cfg, BundleFilter)
-    with open(bundle_fpath, "rb") as src, open(save_dst, "wb") as dst:
-        src.seek(filter_cfg.offset)
-        dst.write(src.read(filter_cfg.len))
+class BundledRecreateFailed(Exception): ...
+
+
+class SlicedRecreateFailed(Exception): ...
+
+
+class CompressedRecreateFailed(Exception): ...
+
+
+class _BundleReadyEventWithRevision:
+    def __init__(self) -> None:
+        self._lock = threading.Lock()
+        self._flag = False
+        self._rev_counter = _counter = itertools.count(0)
+        self._rev = next(_counter)
+
+    def is_set(self) -> tuple[bool, int]:
+        with self._lock:
+            return self._flag, self._rev
+
+    def set(self) -> int:
+        with self._lock:
+            self._flag = True
+            self._rev = next(self._rev_counter)
+            return self._rev
+
+    def clear(self, rev: int) -> None:
+        with self._lock:
+            if rev != self._rev:
+                return
+            self._flag = False
+
+
+SLICE_READ_BUFFER_SIZE = 8 * 1024**2
 
 
 def recreate_sliced_resource(
@@ -52,7 +87,8 @@ def recreate_sliced_resource(
     tmp_save_dst = save_dst.parent / tmp_fname(str(entry.resource_id))
     with open(tmp_save_dst, "wb") as dst:
         for _slice in slices:
-            dst.write(_slice.read_bytes())
+            with open(_slice, "rb") as _slice_f:
+                shutil.copyfileobj(_slice_f, dst, SLICE_READ_BUFFER_SIZE)
     os.replace(tmp_save_dst, save_dst)
 
 
@@ -81,6 +117,12 @@ class ResourceDownloadInfo:
     compressed_origin_size: Optional[int] = None
 
 
+PerBundleLock: TypeAlias = threading.Lock
+BundleRecord: TypeAlias = (
+    "tuple[ResourceTableManifest, PerBundleLock, _BundleReadyEventWithRevision]"
+)
+
+
 class PrepareResourceHelper:
     """Helper for processing resources.
 
@@ -100,7 +142,7 @@ class PrepareResourceHelper:
 
         # for bundled entries, only allow one task to actually rebuild the bundle,
         self._bundle_process_lock = threading.Lock()
-        self._bundle: dict[int, Path] = {}
+        self._bundle_per_locks: dict[int, BundleRecord] = {}
 
         self._thread_local = threading.local()
 
@@ -121,35 +163,89 @@ class PrepareResourceHelper:
             self._thread_local.dctx = dctx
             return dctx
 
+    def _prepare_one_bundle(
+        self, _bundle_f: Path, _bundle_entry: ResourceTableManifest
+    ):
+        # try to re-use already prepared bundle file resource
+        if (
+            _bundle_f.is_file()
+            and file_sha256(_bundle_f).digest() == _bundle_entry.digest
+        ):
+            return
+
+        _bundle_save_tmp = self._download_dir / tmp_fname(
+            str(_bundle_entry.resource_id)
+        )
+        yield from self._prepare_resource(_bundle_entry, _bundle_save_tmp)
+        # unconditionally override the previous file
+        os.replace(_bundle_save_tmp, _bundle_f)
+
     def _prepare_bundled_resource(
         self, entry: ResourceTableManifest, save_dst: Path
     ) -> Generator[ResourceDownloadInfo]:
-        assert isinstance(entry.filter_applied, BundleFilter)
-        # NOTE: prevent the same bundle being prepared again and again
-        bundle_rsid = entry.filter_applied.list_resource_id()
+        filter_cfg = entry.filter_applied
+        assert isinstance(filter_cfg, BundleFilter)
+
+        # NOTE: only do one query for each bundle
+        bundle_rsid = filter_cfg.list_resource_id()
         with self._bundle_process_lock:
-            if bundle_rsid not in self._bundle:
-                bundle_entry = self._orm_pool.orm_select_entry(resource_id=bundle_rsid)
-                logger.debug(f"Requesting bundle({bundle_rsid=}): {bundle_entry}")
-                _bundle_save_tmp = self._download_dir / tmp_fname(str(bundle_rsid))
-                bundle_save_dst = self._download_dir / bundle_entry.digest.hex()
+            if not (_bundle_info := self._bundle_per_locks.get(bundle_rsid)):
+                _bundle_entry = self._orm_pool.orm_select_entry(resource_id=bundle_rsid)
+                assert _bundle_entry, (
+                    f"potential broken OTA image detected, error entry rsid: {bundle_rsid}"
+                )
+                self._bundle_per_locks[bundle_rsid] = (
+                    _bundle_entry,
+                    _bundle_prepare_lock := PerBundleLock(),
+                    _bundle_ready_event := _BundleReadyEventWithRevision(),
+                )
+            else:
+                _bundle_entry, _bundle_prepare_lock, _bundle_ready_event = _bundle_info
 
-                yield from self._prepare_resource(bundle_entry, _bundle_save_tmp)
-                os.replace(_bundle_save_tmp, bundle_save_dst)
-                self._bundle[bundle_rsid] = bundle_save_dst
-            bundle_fpath = self._bundle[bundle_rsid]
+        _bundle_f = self._download_dir / _bundle_entry.digest.hex()
+        # Only trigger upper to download resources when _bundle_ready_event is not set.
+        # To prevent other waiting threads just deadly waiting if the thread that does
+        #   the downloading failed and raised exception, other waiting threads will also
+        #   keep trying to get the _bundle_prepare_lock and prepare bundle.
+        for _ in range(MAX_ITERS_FOR_WAITING_BUNDLE):
+            _is_set, bundle_rev = _bundle_ready_event.is_set()
+            if _is_set:
+                break
 
-        # NOTE: keep the bundle for later use, but if recreate from bundle failed, we delete the bundle
-        #       to let otaclient do the re-downloading.
-        try:
-            recreate_bundled_resource(entry, bundle_fpath, save_dst)
-        except Exception as e:
-            logger.error(
-                f"failed to get resource {entry} from bundle: {e}, remove the bundle!",
-                exc_info=e,
+            if _bundle_prepare_lock.acquire(blocking=False):
+                try:
+                    yield from self._prepare_one_bundle(_bundle_f, _bundle_entry)
+                    bundle_rev = _bundle_ready_event.set()
+                    break
+                except Exception as e:
+                    raise BundledRecreateFailed(
+                        f"failed to prepare bundle {_bundle_entry}: {e!r}"
+                    ) from e
+                finally:
+                    _bundle_prepare_lock.release()
+            time.sleep(WAITING_BUNDLE_INTERVAL)
+        else:
+            raise BundledRecreateFailed(
+                f"timeout waiting for bundle {_bundle_entry=} ready, will retry"
             )
-            bundle_fpath.unlink(missing_ok=True)
-            raise
+
+        # NOTE: keep the bundle for later use, but if recreate from bundle failed,
+        #       we clear the _bundle_ready_event and raise exception to upper to
+        #       trigger a re-downloading of the resources.
+        try:
+            with open(_bundle_f, "rb") as src:
+                src.seek(filter_cfg.offset)
+                _data = src.read(filter_cfg.len)
+
+            if sha256(_data).digest() != entry.digest:
+                raise ValueError(f"digest mismatch when processing {entry=}")
+            with open(save_dst, "wb") as dst:
+                dst.write(_data)
+        except Exception as e:
+            _bundle_ready_event.clear(bundle_rev)
+            raise BundledRecreateFailed(
+                f"failed to extract {entry} from bundle {_bundle_entry}: {e!r}"
+            ) from e
 
     def _prepare_compressed_resources(
         self, entry: ResourceTableManifest, save_dst: Path
@@ -159,34 +255,64 @@ class PrepareResourceHelper:
 
         compressed_rsid = _filter_applied.list_resource_id()
         compressed_entry = self._orm_pool.orm_select_entry(resource_id=compressed_rsid)
+        assert compressed_entry, (
+            f"potential broken OTA image detected, error entry rsid: {compressed_rsid}"
+        )
+        compressed_digest = compressed_entry.digest
 
         # NOTE(20250917): if the compressed entry is not sliced, we directly tell the upper
         #                 caller to decompress on-the-fly during downloading.
         if compressed_entry.filter_applied is None:
             yield ResourceDownloadInfo(
-                digest=compressed_entry.digest,
+                digest=compressed_digest,
                 size=compressed_entry.size,
                 save_dst=save_dst,
                 compression_alg=_filter_applied.compression_alg,
                 compressed_origin_digest=entry.digest,
                 compressed_origin_size=entry.size,
             )
+            return
+
         # if the compressed entry is sliced, we still need to first recover from slices
-        else:
+        compressed_save_dst = self._download_dir / compressed_digest.hex()
+        if (
+            not compressed_save_dst.is_file()
+            or file_sha256(compressed_save_dst).digest() != compressed_digest
+        ):
             _compressed_save_tmp = self._download_dir / tmp_fname(str(compressed_rsid))
-            compressed_save_dst = self._download_dir / compressed_entry.digest.hex()
             yield from self._prepare_resource(compressed_entry, _compressed_save_tmp)
             os.replace(_compressed_save_tmp, compressed_save_dst)
 
-            try:
-                recreate_zstd_compressed_resource(
-                    entry, compressed_save_dst, save_dst, dctx=self._thread_local_dctx
-                )
-            except Exception as e:
-                logger.error(f"failure during decompressing: {entry}: {e}", exc_info=e)
-                raise
-            finally:
-                compressed_save_dst.unlink(missing_ok=True)
+        try:
+            recreate_zstd_compressed_resource(
+                entry,
+                compressed_save_dst,
+                save_dst,
+                dctx=self._thread_local_dctx,
+            )
+        except Exception as e:
+            _err_msg = f"failure during decompressing: {entry}: {e}"
+            raise CompressedRecreateFailed(_err_msg) from e
+        finally:
+            compressed_save_dst.unlink(missing_ok=True)
+
+    def _prepare_one_slice(
+        self, slice_save_dst: Path, slice_entry: ResourceTableManifest
+    ):
+        # try to re-use previously prepared slice
+        if (
+            slice_save_dst.is_file()
+            and file_sha256(slice_save_dst).digest() == slice_entry.digest
+        ):
+            return
+
+        _slice_save_tmp = self._download_dir / tmp_fname(str(slice_entry.resource_id))
+        yield ResourceDownloadInfo(
+            digest=slice_entry.digest,
+            size=slice_entry.size,
+            save_dst=_slice_save_tmp,
+        )
+        os.replace(_slice_save_tmp, slice_save_dst)
 
     def _prepare_sliced_resources(
         self, entry: ResourceTableManifest, save_dst: Path
@@ -196,34 +322,29 @@ class PrepareResourceHelper:
         slices_fpaths: list[Path] = []
         for _slice_rsid in slices_rsid:
             _slice_entry = self._orm_pool.orm_select_entry(resource_id=_slice_rsid)
-            _slice_digest = _slice_entry.digest
-
-            _slice_save_tmp = self._download_dir / tmp_fname(str(_slice_rsid))
-            # NOTE: in case when the slice is shared by multiple resources, we suffix
-            #       the resource_id to the fname.
-            _slice_save_dst = (
-                self._download_dir / f"{_slice_digest.hex()}_{entry.resource_id}"
+            assert _slice_entry, (
+                f"potential broken OTA image detected, error entry rsid: {_slice_rsid}"
             )
-            slices_fpaths.append(_slice_save_dst)
 
             # NOTE: slice SHOULD NOT be filtered again, it MUST be the leaves in the resource
             #       filter applying tree.
             assert _slice_entry.filter_applied is None
-            yield ResourceDownloadInfo(
-                digest=_slice_entry.digest,
-                size=_slice_entry.size,
-                save_dst=_slice_save_tmp,
+
+            # NOTE: in case when the slice is shared by multiple resources, we suffix
+            #       the resource_id to the fname.
+            _slice_save_dst = (
+                self._download_dir / f"{_slice_entry.digest.hex()}_{entry.resource_id}"
             )
-            os.replace(_slice_save_tmp, _slice_save_dst)
+            yield from self._prepare_one_slice(_slice_save_dst, _slice_entry)
+            slices_fpaths.append(_slice_save_dst)
 
         try:
             recreate_sliced_resource(entry, slices_fpaths, save_dst)
         except Exception as e:
-            logger.error(
-                f"failed to recreate sliced resource, remove all slices: {e}",
-                exc_info=e,
+            _err_msg = (
+                f"failed to recreate sliced resource {entry}, remove all slices: {e}"
             )
-            raise
+            raise SlicedRecreateFailed(_err_msg) from e
         finally:
             # after recovering the target resource, cleanup the slices.
             # also, if combination failed, also cleanup the slices to let
@@ -279,3 +400,78 @@ class PrepareResourceHelper:
             )
 
         return target_entry, _gen()
+
+
+SHA256DIGEST_HEX_LEN = 64
+
+
+class ResumeOTADownloadHelper:
+    def __init__(
+        self,
+        download_dir: Path,
+        rst_helper: ResourceTableDBHelper,
+        *,
+        max_concurrent: int,
+        db_conn_num: int = 1,  # serialize accessing
+    ) -> None:
+        self._download_dir = download_dir
+        self._rst_orm_pool = rst_helper.get_orm_pool(db_conn_num)
+        self._se = threading.Semaphore(max_concurrent)
+
+    def _check_one_resource_at_thread(
+        self,
+        _fpath: Path,
+        _digest: bytes,
+        _sliced_target_rsid: str,
+    ) -> None:
+        try:
+            if _sliced_target_rsid and not self._rst_orm_pool.orm_check_entry_exist(
+                resource_id=int(_sliced_target_rsid)
+            ):
+                return remove_file(_fpath)
+
+            if (
+                not self._rst_orm_pool.orm_check_entry_exist(digest=_digest)
+                or file_sha256(_fpath).digest() != _digest
+            ):
+                return remove_file(_fpath)
+        except Exception:
+            remove_file(_fpath)
+        finally:
+            self._se.release()
+
+    def check_download_dir(self) -> int:
+        """Scan through OTA download dir and try to recover resources."""
+        _count = 0
+        with ThreadPoolExecutor(
+            thread_name_prefix="resume_ota_download"
+        ) as pool, os.scandir(self._download_dir) as it:
+            for entry in it:
+                if (
+                    not entry.is_file(follow_symlinks=False)
+                    or len(entry.name) < SHA256DIGEST_HEX_LEN
+                    or entry.name.startswith("tmp")
+                ):
+                    remove_file(Path(entry.path))
+                    continue
+
+                entry_fname = entry.name
+                # NOTE: for slice, a suffix will be appended to the filename.
+                _digest_hex = entry_fname[:SHA256DIGEST_HEX_LEN]
+                # see L330-L332, a slice will be named as <slice_digest>_<target_resource_id>
+                _sliced_target_rsid = entry_fname[SHA256DIGEST_HEX_LEN + 1 :]
+                try:
+                    _digest = bytes.fromhex(_digest_hex)
+                except Exception:
+                    remove_file(Path(entry.path))
+                    continue
+
+                self._se.acquire()
+                _count += 1
+                pool.submit(
+                    self._check_one_resource_at_thread,
+                    Path(entry.path),
+                    _digest,
+                    _sliced_target_rsid,
+                )
+        return _count
